2 py files were united to the prototype
