+++ cruxbot/cruxbot.py
@@ -294,6 +294,14 @@ def bot_liking_posts(api,  # type: Api
 #----------------- END BOT FUNCTIONS -----------------#
 #-----------------------------------------------------#
 
+def display_menu_fb():
+    print("1.Dar like a un posteo\n2.Leer posteo\n3.Subir posteo o foto\n4.Actualizar perfil\n5.Listar amigos\n6.Listar seguidores\n7.Seguir a un usuario\n8.Enviar un mensaje a un usuario")
+
+    
+def display_menu_ig():
+    print("1.Buscar usuario\n2.Listar seguidores\n3.Modificar publicacion\n4.Publicar foto\n5.Mostrar posteo")
+
+
 def main():
     api = Api(
         app_id = APP_ID,
